Move ed25519_verify to crypto crate
