+++ packages/vm/src/lib.rs
@@ -6,7 +6,6 @@ mod calls;
 mod checksum;
 mod compatibility;
 mod conversion;
-mod crypto;
 mod environment;
 mod errors;
 mod features;
